requirements.txt
annotated-types==0.7.0
anyio==4.9.0
botocore==1.38.36
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.115.12
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.86.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
